requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.0
openai==1.12.0
python-dotenv==1.0.1
openpyxl==3.1.2 